GRAILS-9528 - remove GrailsMock and mockFor method
