+++ grails-test-suite-uber/src/test/groovy/grails/test/mixin/FiltersUnitTestMixinTests.groovy
@@ -1,7 +1,6 @@
 package grails.test.mixin
 
 import grails.artefact.Artefact
-import grails.test.GrailsMock
 import grails.test.mixin.web.FiltersUnitTestMixin
 
 import org.junit.After
@@ -13,7 +12,7 @@ import org.springframework.beans.factory.config.MethodInvokingFactoryBean
 class FiltersUnitTestMixinTests {
 
     AuthorController controller
-    GrailsMock autowiredServiceMock
+    AutowiredService autowiredService
 
     @Before
     void setUp() {
@@ -89,7 +88,7 @@ class FiltersUnitTestMixinTests {
         defineBeans {
             autowiredService(MethodInvokingFactoryBean) {
                 targetObject = this
-                targetMethod = 'mockAutowiredService'
+                targetMethod = 'setupService'
             }
         }
         mockFilters(AutowiredFilters)
@@ -98,7 +97,7 @@ class FiltersUnitTestMixinTests {
             controller.list()
         }
 
-        autowiredServiceMock.verify()
+        assert 1 == autowiredService.sessionSetupCounter
     }
 
     @Test
@@ -107,7 +106,7 @@ class FiltersUnitTestMixinTests {
         defineBeans {
             autowiredService(MethodInvokingFactoryBean) {
                 targetObject = this
-                targetMethod = 'mockAutowiredService'
+                targetMethod = 'setupService'
             }
         }
 
@@ -115,13 +114,11 @@ class FiltersUnitTestMixinTests {
             controller.list()
         }
 
-        autowiredServiceMock.verify()
+        assert 1 == autowiredService.sessionSetupCounter
     }
 
-    AutowiredService mockAutowiredService() {
-        this.autowiredServiceMock = mockFor(AutowiredService)
-        this.autowiredServiceMock.demand.setupSession(1) {}
-        return this.autowiredServiceMock.createMock()
+    AutowiredService setupService() {
+        this.autowiredService = new AutowiredService()
     }
 }
 
@@ -192,6 +189,8 @@ class AutowiredFilters {
 }
 
 class AutowiredService {
-
-    void setupSession() {}
+    int sessionSetupCounter = 0
+    void setupSession() {
+        sessionSetupCounter++
+    }
 }
