+++ grails-test-suite-uber/src/test/groovy/grails/test/mixin/MyService.groovy
@@ -1,0 +1,7 @@
+package grails.test.mixin
+
+class MyService {
+    int identity() {
+        System.identityHashCode(this)
+    }
+}
