--- grails-test-suite-uber/src/test/groovy/grails/test/mixin/ServiceAndMockForTests.groovy
@@ -1,56 +1,0 @@
-package grails.test.mixin
-
-import org.junit.Test
-
-@TestFor(MyService)
-class ServiceAndMockForTests {
-
-    @Test
-    void testProvaUsingMockFor() {
-        def mockForUtilsControl = mockFor(MockForUtils2)
-        mockForUtilsControl.demand.isMockForWorking(1..2) {->
-            return true
-        }
-        service.utils = mockForUtilsControl.createMock()
-        service.prova()
-
-        assert service.prova()
-
-        // run validation on mocks
-        mockForUtilsControl.verify()
-    }
-
-    @Test
-    void testProvaUsingMockForAgain() {
-        def mockForUtilsControl = mockFor(MockForUtils2)
-        mockForUtilsControl.demand.isMockForWorking(1..1) { -> true }
-        service.utils = mockForUtilsControl.createMock()
-        assert service.prova()
-
-        // run validation on mocks
-        mockForUtilsControl.verify()
-    }
-
-    @Test
-    void testIndexWithoutUsingMockForAgain() {
-        service.prova()
-    }
-
-}
-
-class MyService {
-    def utils
-    boolean prova() {
-        utils?.isMockForWorking()
-    }
-    
-    int identity() {
-        System.identityHashCode(this)
-    }
-}
-
-class MockForUtils2 {
-    boolean isMockForWorking() {
-        return false
-    }
-}
